Pitstop/consumer/requirements.txt
kafka-python
pandas
scikit-learn
flask
pillow
